+++ contao/dca/tl_block.php
@@ -49,38 +49,12 @@ $GLOBALS['TL_DCA']['tl_block'] = [
             ],
         ],
         'operations'        => [
-            'edit'       => [
-                'label' => &$GLOBALS['TL_LANG']['tl_block']['edit'],
-                'href'  => 'table=tl_block_module',
-                'icon'  => 'edit.gif',
-            ],
-            'editHeader' => [
-                'label' => &$GLOBALS['TL_LANG']['tl_block']['editHeader'],
-                'href'  => 'act=edit',
-                'icon'  => 'header.gif',
-            ],
-            'copy'       => [
-                'label' => &$GLOBALS['TL_LANG']['tl_block']['copy'],
-                'href'  => 'act=copy',
-                'icon'  => 'copy.gif',
-            ],
-            'delete'     => [
-                'label'      => &$GLOBALS['TL_LANG']['tl_block']['delete'],
-                'href'       => 'act=delete',
-                'icon'       => 'delete.gif',
-                'attributes' => 'onclick="if (!confirm(\'' . ($GLOBALS['TL_LANG']['MSC']['deleteConfirm'] ?? null) . '\')) return false; Backend.getScrollOffset();"',
-            ],
-            'toggle'     => [
-                'label'           => &$GLOBALS['TL_LANG']['tl_block']['toggle'],
-                'icon'            => 'visible.gif',
-                'attributes'      => 'onclick="Backend.getScrollOffset();return AjaxRequest.toggleVisibility(this,%s)"',
-                'button_callback' => [BlockContainer::class, 'toggleIcon'],
-            ],
-            'show'       => [
-                'label' => &$GLOBALS['TL_LANG']['tl_block']['show'],
-                'href'  => 'act=show',
-                'icon'  => 'show.gif',
-            ],
+            'edit',
+            'children',
+            'copy',
+            'delete',
+            'toggle',
+            'show',
         ],
     ],
     // Palettes
@@ -131,9 +105,9 @@ $GLOBALS['TL_DCA']['tl_block'] = [
             'sql'       => "varchar(255) NOT NULL default ''",
         ],
         'published'  => [
-            'label'     => &$GLOBALS['TL_LANG']['tl_block']['published'],
             'exclude'   => true,
             'filter'    => true,
+            'toggle'    => true,
             'inputType' => 'checkbox',
             'eval'      => ['doNotCopy' => true, 'submitOnChange' => true],
             'sql'       => "char(1) NOT NULL default '1'",